commit2

+++ TemperatureConverter/Program.cs
@@ -8,7 +8,7 @@ class Program
 
     static void Main(string[] args)
     {
-        int Temperature;
+        decimal Temperature;
 
 
 
@@ -18,15 +18,11 @@ class Program
             try
             {
                 Console.WriteLine("Enter Temperature in C:");
-                Temperature = Convert.ToInt32(Console.ReadLine());
-                Console.WriteLine(Temperature * 9 / 5 + 32);
+                Temperature = Convert.ToDecimal(Console.ReadLine());
+                Console.WriteLine( Temperature * 9 / 5 + 32);
                 Console.WriteLine("Enter Temperature in F: ");
-                Temperature = Convert.ToInt32(Console.ReadLine());
+                Temperature = Convert.ToDecimal(Console.ReadLine());
                 Console.WriteLine((Temperature - 32) * 5 / 9);
-
-
-
-
             }
             catch
             {
